+++ x/oracle/abci.go
@@ -10,8 +10,7 @@ import (
 // BeginBlocker re-calculates and saves the rolling seed value based on block hashes.
 func BeginBlocker(ctx sdk.Context, k keeper.Keeper) error {
 	// Reward a portion of block rewards (inflation + tx fee) to active oracle validators.
-	k.AllocateTokens(ctx, ctx.VoteInfos())
-	return nil
+	return k.AllocateTokens(ctx, ctx.VoteInfos())
 }
 
 // EndBlocker cleans up the state during end block. See comment in the implementation!